semi-fixed
